Webdev Branch Init Sync
This is the first branch sync. The only difference between this sync and
the one on the master is the introduction to the "colors" module, and a
small express error fix. Make sure you save your custom batch file
somewhere other than the repository so no one has to sync to it.