Merged Play/Pause into Toggle, Prettied up
Just a couple of changes. Made a more visible header, created a toggle
between play and pause on the same button, and added a nice picture of
space so we don't have to stare at a white void anymore. Don't forget to
replace the test.mp3! Your new mp3 needs to be titled the same thing.