Skip Button w/ Test Playlist
Only 3 are currently loaded in the fixated playlist, but this is the
log/songs that were supplied by Boston.
There are still some issues. Once all the songs have played, you can't
use the restart button. The restart button also does not let you listen
to previous songs (if we wanted to implement that)
I'll continue to stomp out the bugs and see if I can get tag info up.
Sadly, this will also have to be hard-coded in, but I'm feeling pretty
confident will have a lot of this done by Thursday.