Big Change Read Description
Change can be scary but hopefully this change will help. I moved all the
onclick properties to the playerScript.js. This is should be more robust
and make the play run smoother I also moved methods around to allow them
to be grouped based on use. I changed some methods to accommodate some
other change. This should help us to reach our goal easier and all
that's left is the playlist implementation